app/requirements.txt
gunicorn==20.0.4
Flask==2.0.1
numpy==1.19.5
opencv-python==4.5.3.56
tensorflow-cpu==2.5.0